new functions

+++ build.bat
@@ -1,15 +1,18 @@
 @echo off
-echo Установка PyInstaller...
-.\venv\Scripts\python.exe -m pip install pyinstaller
+echo Обновление библиотек...
+python -m pip install --upgrade pip
+python -m pip install --upgrade pydantic aiogram
+python -m pip install -r requirements.txt
+python -m pip install pyinstaller
 
 echo Сборка проекта...
-.\venv\Scripts\pyinstaller.exe --onefile --name "PC_Control_Bot" main.py
+python -m PyInstaller --onefile --noconsole --noconfirm --name "PC_Control_Bot" main.py
 
 echo Очистка временных файлов сборки...
-move /y dist\PC_Control_Bot.exe .\PC_Control_Bot.exe
-rmdir /s /q build
-rmdir /s /q dist
-del /q PC_Control_Bot.spec
+if exist dist\PC_Control_Bot.exe move /y dist\PC_Control_Bot.exe .\PC_Control_Bot.exe
+if exist build rmdir /s /q build
+if exist dist rmdir /s /q dist
+if exist PC_Control_Bot.spec del /q PC_Control_Bot.spec
 
 echo Сборка завершена! Готовый файл: PC_Control_Bot.exe
 pause
